\name{mat.plotjump}
\alias{mat.plotjump}
 
\title{Plot Jump Analysis}
\description{
  Plots the result of mat.jump for visualization of optimal alpha value.
}
\usage{
mat.plotjump(inJumpObj)
}
 
\arguments{
  \item{inJumpObj}{List: any output from applying mat.jump function.}
}
\details{
  This is another simple plotting function useful for study and modification by users.
}
\value{
A graphical device with a plot.
}
\references{SAWADA, M.  2006. An open-source implementation of the modern analog technique (MAT) within the R computing environment. Computers and Geosciences, Vol 32(6):818-833, DOI information: 10.1016/j.cageo.2005.10.008}
\author{Professor Michael Sawada}
 

\seealso{\code{\link{mat.jumprecon}}, \code{\link{mat.jump}},\code{\link{mat.dissim}}, }
 
\keyword{ plotting } 
 

 

