\name{mat.roc.up}
\alias{mat.roc.up}
 
\title{ROC Analysis
 
}
\description{TBD
 
}
\usage{
mat.roc.up(mattools.roclist, rocEvalSeq = seq(0, 2, 0.01))
}
 
\arguments{
  \item{mattools.roclist}{TBD
 
}
  \item{rocEvalSeq}{TBD
 
}
}
\details{TBD
 
}
\value{TBD
 
}
\references{SAWADA, M.  2006. An open-source implementation of the modern analog technique (MAT) within the R computing environment. Computers and Geosciences, Vol 32(6):818-833, DOI information: 10.1016/j.cageo.2005.10.008}
\author{ Dr. M. Sawada}
 
 
\keyword{ functions }
 
