\name{mat.ROCcalc}
\alias{mat.ROCcalc}
 
\title{ ROC calculation
 
}
\description{ Needs docs
 
}
\usage{
mat.ROCcalc(truth, data, evalseq, method = "trap")
}
 
\arguments{
  \item{truth}{ TBD
 
}
  \item{data}{ TBD
 
}
  \item{evalseq}{TBD
 
}
  \item{method}{TBD
 
}
}
\details{TBD
 
}
\value{
TBD
}
\references{SAWADA, M.  2006. An open-source implementation of the modern analog technique (MAT) within the R computing environment. Computers and Geosciences, Vol 32(6):818-833, DOI information: 10.1016/j.cageo.2005.10.008}
\author{ Dr. M. Sawada}
 
 
\keyword{ functions }
 
