\name{mattools-package}
\alias{mattools-package}
\alias{mattools}
\docType{package}
\title{Modern Calibration Functions for the Modern Analog Technique
 
}
\description{This package in includes functions for receiver operating
        characteristic (ROC) analyses as well as Monte Carlo
        simulation.  It includes specific graphical functions for
        interpreting the output of these techniques.
 
}
\details{
\tabular{ll}{
Package: \tab mattools\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2012-07-23\cr
License: \tab GPL 2.0\cr
}
 
}
\author{Dr. M. Sawada.
 

Maintainer: steven mosher <moshersteven@gmail.com>
 
}
\references{
SAWADA, M.  2006. An open-source implementation of the modern analog technique (MAT) within the R computing environment. Computers and Geosciences, Vol 32(6):818-833, DOI information: 10.1016/j.cageo.2005.10.008
}
 
\keyword{ package }
 
 
