\name{polyarea}
\alias{polyarea}
 
\title{calculate area of polygon
 
}
\description{TBD
 
}
\usage{
polyarea(inpolygon)
}
 
\arguments{
  \item{inpolygon}{TBD
 
}
}
\details{
TBD
}
\value{
 TBD
}
 
\references{SAWADA, M.  2006. An open-source implementation of the modern analog technique (MAT) within the R computing environment. Computers and Geosciences, Vol 32(6):818-833, DOI information: 10.1016/j.cageo.2005.10.008}
\author{ Dr. M. Sawada}
 

 
\keyword{ internal }
 
