\name{spherical.direction.f}
\alias{spherical.direction.f}
 
\title{spherical direction
 
}
\description{TBD
 
}
\usage{
spherical.direction.f(cLonLat, oLonLat)
}
 
\arguments{
  \item{cLonLat}{TBD
 
}
  \item{oLonLat}{TBD
 
}
}
\details{TBD
 
}
\value{TBD
 
}
 
\references{SAWADA, M.  2006. An open-source implementation of the modern analog technique (MAT) within the R computing environment. Computers and Geosciences, Vol 32(6):818-833, DOI information: 10.1016/j.cageo.2005.10.008}
\author{ Dr. M. Sawada}

 
\keyword{ internal }
 
